tests/wpt/tests/mathml/relations/html5-tree/a-hreflang-getter.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MathMLAnchorElement.hreflang getter</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <math>
      <a id="test" href="a" hreflang="en">Example Link</a>
    </math>
  </body>
  <script>
      test(function() {
          var a = document.getElementById("test");
          assert_equals(a.hreflang, "en");
      }, "Test anchor's hreflang getter");
  </script>
</html>
